refactor(web): fix lib->features layering inversion
- Move sanitizeAuthRedirect (pure same-origin redirect whitelist) from
  features/auth/lib/auth-redirect.ts to src/lib/helpers/
  sanitize-auth-redirect.ts; http-client.ts now imports inside lib, and
  features/auth + routes/sign-in import downward from lib. The lib-level
  401 sign-in redirect keeps the identical open-redirect whitelist.
- Relocate search-params-resilience.test.ts from src/lib/helpers/__tests__
  to src/__tests__: it exercises feature schemas, which made src/lib depend
  on @/features in the test domain.
- grep -rn '@/features' src/lib/ now returns nothing.
- Register the four components/layout -> features nav-metadata edges as
  documented exceptions in the new docs/internal/web-package-boundaries.md
  (layering rules + exit criterion), since inverting the drill-in sidebar
  registry is too invasive for this wave.
